+++ src/main/java/com/pfchoice/springboot/repositories/LeadStatusRepository.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Repository;
 import com.pfchoice.springboot.model.LeadStatus;
 
 @Repository
-public interface LeadStatusRepository extends JpaRepository<LeadStatus, Byte> {
+public interface LeadStatusRepository extends JpaRepository<LeadStatus, Short> {
 
-	public LeadStatus findById(Byte id);
+	public LeadStatus findById(Short id);
 	
 	public LeadStatus findByDescription(String description);
